# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getCurrentAppUser
      tags: [auth]
      summary: Get the current signed-in application user
      responses:
        "200":
          description: Current application user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentAppUserResponse"
        "401":
          description: Sign-in required
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in with a provisioned application account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Signed-in application user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentAppUserResponse"
        "401":
          description: Invalid credentials
        "503":
          description: Authentication unavailable
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List application users and roles
      responses:
        "200":
          description: Users available for role management
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersResponse"
        "403":
          description: Administrator access required
    post:
      operationId: createAdminUser
      tags: [admin]
      summary: Create a username and password account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProvisionedUserInput"
      responses:
        "201":
          description: Created application user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentAppUserResponse"
        "400":
          description: Invalid account details
        "403":
          description: Administrator access required
        "409":
          description: Username is already in use or reserved
  /admin/users/{userId}/role:
    patch:
      operationId: updateAdminUserRole
      tags: [admin]
      summary: Update an application's user role
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppUserRoleUpdate"
      responses:
        "200":
          description: Updated application user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentAppUserResponse"
        "400":
          description: Invalid role
        "403":
          description: Administrator access required
        "404":
          description: User not found
        "409":
          description: The last administrator cannot be demoted
  /admin/audit-trails:
    get:
      operationId: listAuditTrails
      tags: [admin]
      summary: List recent audit trail events
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: number
            minimum: 1
            maximum: 200
            default: 100
        - name: offset
          in: query
          required: false
          schema:
            type: number
            minimum: 0
            default: 0
      responses:
        "200":
          description: Audit events ordered newest first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditTrailsResponse"
        "403":
          description: Administrator access required
...
components:
  schemas:
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
...
    AppRole:
      type: string
      enum: [admin, manager, viewer]
    AppUser:
      type: object
      properties:
        userId:
          type: string
        email:
          type: string
          description: Contact email, or a local placeholder when omitted
        fullName:
          type: string
        username:
          type: ["string", "null"]
          description: Login username for locally provisioned accounts
        role:
          $ref: "#/components/schemas/AppRole"
        createdAt:
          type: string
          format: date-time
        lastSeenAt:
          type: ["string", "null"]
          format: date-time
      required: [userId, email, fullName, username, role, createdAt, lastSeenAt]
    CurrentAppUserResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AppUser"
      required: [user]
    AdminUsersResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AppUser"
      required: [users]
    AppUserRoleUpdate:
      type: object
      properties:
        role:
          $ref: "#/components/schemas/AppRole"
      required: [role]
    ProvisionedUserInput:
      type: object
      properties:
        fullName:
          type: string
          minLength: 1
          maxLength: 200
        username:
          type: string
          minLength: 3
          maxLength: 64
          pattern: "^[A-Za-z0-9][A-Za-z0-9._-]{2,63}$"
        email:
          type: string
          maxLength: 320
          description: Optional contact email; it is not used to sign in or verified.
        password:
          type: string
          minLength: 1
          maxLength: 1024
        role:
          $ref: "#/components/schemas/AppRole"
      required: [fullName, username, password, role]
    AuditTrailEntry:
      type: object
      properties:
        id:
          type: number
        actorUserId:
          type: string
        actorEmail:
          type: ["string", "null"]
        actorRole:
          $ref: "#/components/schemas/AppRole"
        method:
          type: string
        path:
          type: string
        statusCode:
          type: number
        requestId:
          type: ["string", "null"]
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - actorUserId
        - actorEmail
        - actorRole
        - method
        - path
        - statusCode
        - requestId
        - metadata
        - createdAt
    AuditTrailsResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/AuditTrailEntry"
        total:
          type: number
        limit:
          type: number
        offset:
          type: number
      required: [events, total, limit, offset]
...