# lib/api-zod/src/generated/types/appUser.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AppRole } from './appRole';

export interface AppUser {
  userId: string;
  /** Contact email, or a local placeholder when omitted */
  email: string;
  fullName: string;
  /**
     * Login username for locally provisioned accounts
     * @nullable
     */
  username: string | null;
  role: AppRole;
  createdAt: Date;
  /** @nullable */
  lastSeenAt: Date | null;
}