# lib/api-zod/src/generated/types/provisionedUserInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AppRole } from './appRole';

export interface ProvisionedUserInput {
  /**
     * @minLength 1
     * @maxLength 200
     */
  fullName: string;
  /**
     * @minLength 3
     * @maxLength 64
     * @pattern ^[A-Za-z0-9][A-Za-z0-9._-]{2,63}$
     */
  username: string;
  /**
     * Optional contact email; it is not used to sign in or verified.
     * @maxLength 320
     */
  email?: string;
  /**
     * @minLength 1
     * @maxLength 1024
     */
  password: string;
  role: AppRole;
}